requirements.txt
numpy
opencv-python
mediapipe
protobuf==3.20.*
RPi.GPIO 
python-dotenv
